Define WINVER
In some cases with MinGW-w64 on Windows, the build can fail because
"'WC_ERR_INVALID_CHARS' was not declared in this scope".

The issue has been seen using the 32-bit version of MinGW-w64 from
WinLibs when compiling the core library via Go with GOARCH=386.

To improve compatibility we can define WINVER as 0x0600 (Vista).