Control stopping the run loop when last window is closed (macOS)
We used to let the system/UI framework automatically handle when to
terminate the app after the last window was closed. We used custom logic
to stop the run loop instead of terminating the whole process.

Now we handle the stopping of the run loop without having to awkwardly
initiate termination just to block it and do something else instead.

This can allow us to e.g. conditionally stop the run loop when a user the
library is responsible for their own run loop.